refactor: UserException 을 user/domain 으로 이동, require/check → 커스텀 예외 교체
도메인이 자기방어하도록 User.promoteToMember, updateNickname 의 require/check 를
UserException 커스텀 예외로 교체. UserException 을 service → domain 패키지로 이동해
domain → service 레이어 역방향 의존 제거.
